services/access-agent/requirements.txt
# Gough Access Agent Dependencies
# Python 3.13+ required

# SSH server implementation
paramiko>=3.4.0

# HTTP client for management server API
httpx>=0.27.0

# JWT token handling
pyjwt>=2.8.0

# Configuration management
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Cryptography for key handling
cryptography>=42.0.0

# Logging and monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# Process management
psutil>=5.9.0

# Penguin shared libraries (published packages)
penguin-libs>=0.1.0
penguin-licensing>=0.1.0
